frontEnd/getMetrics/requirements.txt
google-cloud-bigquery==2.34.0
functions-framework==3.2.0
Flask==2.3.2
